system component dependencies, system messages, more porting
