+++ tweey/Components/HeadingComponent.cs
@@ -1,0 +1,7 @@
+﻿namespace Tweey.Components;
+
+[EcsComponent]
+struct HeadingComponent
+{
+    public float Heading { get; set; }
+}
